AgriTool-Document-Schema-Extractor/requirements.txt
# Core dependencies for document schema extraction
supabase>=2.0.0
aiohttp>=3.8.0
aiofiles>=23.0.0

# Document processing libraries
PyPDF2>=3.0.0
python-docx>=0.8.11
openpyxl>=3.1.0

# OCR support for scanned documents
pytesseract>=0.3.10
Pillow>=10.0.0

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0